+++ extensions/easydict/src/detectLanauge/detect.ts
@@ -2,12 +2,13 @@
  * @author: tisfeng
  * @createTime: 2022-06-24 17:07
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-10-02 15:50
+ * @lastEditTime: 2022-10-13 00:21
  * @fileName: detect.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
+import { autoDetectLanguageItem, chineseLanguageItem, englishLanguageItem } from "../language/consts";
 import { isValidLangCode } from "../language/languages";
 import { myPreferences } from "../preferences";
 import { appleLanguageDetect } from "../scripts";
@@ -17,7 +18,6 @@ import { hasTencentAppKey, tencentDetect } from "../translation/tencent";
 import { volcanoDetect } from "../translation/volcano/volcanoAPI";
 import { hasVolcanoAppKey } from "../translation/volcano/volcanoSign";
 import { RequestErrorInfo } from "../types";
-import { autoDetectLanguageItem, chineseLanguageItem, englishLanguageItem } from "./../language/consts";
 import { francLangaugeDetect } from "./franc";
 import { DetectedLangModel, LanguageDetectType } from "./types";
 import {
@@ -124,7 +124,7 @@ function raceDetectTextLanguage(lowerCaseText: string): Promise<DetectedLangMode
         .finally(() => {
           detectCount += 1;
           // If the last detection action is still not resolve, return undefined.
-          if (detectCount === detectActionList.length && hasDetectFinished === false) {
+          if (detectCount === detectActionList.length && !hasDetectFinished) {
             console.warn(`last detect action fail, return undefine`);
             resolve(undefined);
           }
@@ -165,11 +165,9 @@ function handleDetectedLanguage(detectedLangModel: DetectedLangModel): Promise<D
     if (!isValidLangCode(detectedLangCode)) {
       return resolve(undefined);
     }
-
     // 2. Iterate API detected language List, check if has detected >= `two` identical valid language.
     const detectedIdenticalLanguages: DetectedLangModel[] = [];
     const detectedTypes: string[] = [];
-
     console.log(`detectedLangCode: ${detectedLangCode}, detectedList: ${apiDetectedLanguageList.length}`);
 
     for (const lang of apiDetectedLanguageList) {
@@ -190,7 +188,7 @@ function handleDetectedLanguage(detectedLangModel: DetectedLangModel): Promise<D
         if (
           (containBingDetect || containBaiduDetect || confirmVolcanoDetect) &&
           isPreferredLanguage(detectedLangCode) &&
-          myPreferences.enableLanguageDetectionSpeedFirst
+          myPreferences.enableDetectLanguageSpeedFirst
         ) {
           detectedLangModel.confirmed = true;
           console.warn(`---> Speed first, API detected 'two' identical 'preferred' language: ${detectedTypes}`);
@@ -298,14 +296,14 @@ function getLocalTextLanguageDetectResult(
         console.log(
           `franc detect preferred but unconfirmed language: ${languageId}, confidence: ${confidence} (>${lowConfidence})`
         );
-        const lowConfidenceDetectTypeResult: DetectedLangModel = {
+        const lowConfidenceDetect: DetectedLangModel = {
           type: francDetectResult.type,
           sourceLangCode: francDetectResult.sourceLangCode,
           youdaoLangCode: languageId,
           confirmed: false,
           detectedLanguageArray: francDetectResult.detectedLanguageArray,
         };
-        return lowConfidenceDetectTypeResult;
+        return lowConfidenceDetect;
       }
     }
   }